Ignore expired routes in `netstat` on macOS
When routes expire in macOS, `netstat -r` shows a trailing exclamation point to indicate this. In the current version of this code, this will set `DEF_GW_IF` to contain a `!`, causing subsequent DNS modifications on OSX to fail.

This change ignores the the `!` that can be set in this case.